+++ composer.json
@@ -1,4 +1,4 @@
-{
+    {
     "name": "symfony/framework-standard-edition",
     "license": "MIT",
     "type": "project",
