+++ src/AppBundle/Tests/Controller/WorkContributeControllerTest.php
@@ -1,10 +1,8 @@
 <?php
 
-namespace Acme\BlogBundle\Tests\Controller;
+namespace AppBundle\Tests\Controller;
 
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
-
-class WorkContributeControllerTest extends WebTestCase
+class WorkContributeControllerTest extends AbstractApiTest
 {
     public function testGet()
     {
@@ -14,17 +12,5 @@ class WorkContributeControllerTest extends WebTestCase
         $response = $client->getResponse();
 
         $this->assertJsonResponse($response, 200);
-    }
-
-    protected function assertJsonResponse($response, $statusCode = 200)
-    {
-        $this->assertEquals(
-            $statusCode, $response->getStatusCode(),
-            $response->getContent()
-        );
-        $this->assertTrue(
-            $response->headers->contains('Content-Type', 'application/json'),
-            $response->headers
-        );
     }
 }
